requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
sqlalchemy
pymysql
werkzeug
python-dotenv
gunicorn
flasgger
PyYAML
